opal: remove 32-bit platforms from atomics
These platforms are losing support with each passing day and it
is time to let them go. They will continue to work with a C11
compiler but Open MPI will no longer maintain ASM for these
platforms.

(cherry picked from commit 6abd667325aa31ceb6dae3e49ceac47fc997ea75)
