+++ opal/include/opal/sys/arm64/timer.h
@@ -6,6 +6,7 @@
  * Copyright (c) 2016      Broadcom Limited. All rights reserved.
  * Copyright (c) 2016      Los Alamos National Security, LLC. All rights
  *                         reserved.
+ * Copyright (c) 2021      Google, LLC. All rights reserved.
  * $COPYRIGHT$
  *
  * Additional copyrights may follow
@@ -16,14 +17,22 @@
 #ifndef OPAL_SYS_ARCH_TIMER_H
 #define OPAL_SYS_ARCH_TIMER_H 1
 
+#if OPAL_ASSEMBLY_ARCH == OPAL_ARM64
 typedef uint64_t opal_timer_t;
+#else
+typedef uint32_t opal_timer_t;
+#endif
 
 static inline opal_timer_t opal_sys_timer_get_cycles(void)
 {
     opal_timer_t ret;
 
     __asm__ __volatile__("isb" ::: "memory");
+#if OPAL_ASSEMBLY_ARCH == OPAL_ARM64
     __asm__ __volatile__("mrs %0,  CNTVCT_EL0" : "=r"(ret));
+#else
+    __asm__ __volatile__("mrs %0,  CNTVCT" : "=r"(ret));
+#endif
 
     return ret;
 }
@@ -31,7 +40,11 @@ static inline opal_timer_t opal_sys_timer_get_cycles(void)
 static inline opal_timer_t opal_sys_timer_get_freq(void)
 {
     opal_timer_t freq;
+#if OPAL_ASSEMBLY_ARCH == OPAL_ARM64
     __asm__ __volatile__("mrs %0,  CNTFRQ_EL0" : "=r"(freq));
+#else
+    __asm__ __volatile__("mrs %0,  CNTFRQ" : "=r"(freq));
+#endif
     return (opal_timer_t)(freq);
 }
