+++ opal/include/opal/sys/timer.h
@@ -72,17 +72,11 @@ BEGIN_C_DECLS
 
 #if defined(DOXYGEN)
 /* don't include system-level gorp when generating doxygen files */
-#elif OPAL_ASSEMBLY_ARCH == OPAL_X86_64
+#elif OPAL_ASSEMBLY_ARCH == OPAL_X86_64 || OPAL_ASSEMBLY_ARCH == OPAL_IA32
 #    include "opal/sys/x86_64/timer.h"
-#elif OPAL_ASSEMBLY_ARCH == OPAL_ARM
-#    include "opal/sys/arm/timer.h"
-#elif OPAL_ASSEMBLY_ARCH == OPAL_ARM64
+#elif OPAL_ASSEMBLY_ARCH == OPAL_ARM64 || OPAL_ASSEMBLY_ARCH == OPAL_ARM
 #    include "opal/sys/arm64/timer.h"
-#elif OPAL_ASSEMBLY_ARCH == OPAL_IA32
-#    include "opal/sys/ia32/timer.h"
-#elif OPAL_ASSEMBLY_ARCH == OPAL_POWERPC32
-#    include "opal/sys/powerpc/timer.h"
-#elif OPAL_ASSEMBLY_ARCH == OPAL_POWERPC64
+#elif OPAL_ASSEMBLY_ARCH == OPAL_POWERPC64 || OPAL_ASSEMBLY_ARCH == OPAL_POWERPC32
 #    include "opal/sys/powerpc/timer.h"
 #endif
